fix JUnit tests - we're returning longs rather than ints now

+++ crux-test/test/crux/api/TestUtils.java
@@ -34,7 +34,7 @@ public class TestUtils {
     static final String documentId = "myDoc";
     static final String versionId = "version";
 
-    static CruxDocument testDocument(int version) {
+    static CruxDocument testDocument(long version) {
         CruxDocument document = CruxDocument.create(documentId);
         return document.plus(versionId, version);
     }
